chore: lint

+++ test/Merkle.t.sol
@@ -16,7 +16,7 @@ contract MerkleTest is Test {
 
     function testInsert() public {
         bytes32 node = keccak256("test");
-        
+
         simpleMerkle.insert(node);
         fullMerkle.insert(node);
 
@@ -35,10 +35,10 @@ contract MerkleTest is Test {
     function testMultipleInserts() public {
         bytes32[] memory nodes = new bytes32[](3);
         nodes[0] = keccak256("test1");
-        nodes[1] = keccak256("test2"); 
+        nodes[1] = keccak256("test2");
         nodes[2] = keccak256("test3");
 
-        for(uint i = 0; i < nodes.length; i++) {
+        for (uint256 i = 0; i < nodes.length; i++) {
             simpleMerkle.insert(nodes[i]);
             fullMerkle.insert(nodes[i]);
 
@@ -53,12 +53,12 @@ contract MerkleTest is Test {
 
     function testVerify() public {
         bytes32 node = keccak256("test");
-        
+
         simpleMerkle.insert(node);
         fullMerkle.insert(node);
 
         bytes32[16] memory proof = fullMerkle.getProof(0);
-        
+
         assertTrue(simpleMerkle.verify(node, proof, 0));
         assertTrue(fullMerkle.verify(node, proof, 0));
         assertEq(simpleMerkle.root(), fullMerkle.root());
@@ -67,24 +67,26 @@ contract MerkleTest is Test {
     function testUpdate() public {
         bytes32 oldNode = keccak256("test");
         bytes32 newNode = keccak256("updated");
-        
+
         simpleMerkle.insert(oldNode);
         fullMerkle.insert(oldNode);
 
         bytes32[16] memory proof = fullMerkle.getProof(0);
-        
+
         simpleMerkle.update(newNode, oldNode, proof, 0);
         fullMerkle.update(newNode, 0);
 
         assertEq(simpleMerkle.root(), fullMerkle.root());
-        
+
         // Verify new node
         proof = fullMerkle.getProof(0);
         assertTrue(simpleMerkle.verify(newNode, proof, 0));
         assertTrue(fullMerkle.verify(newNode, proof, 0));
     }
 
-    function testFuzzInsert(bytes32 node) public {
+    function testFuzzInsert(
+        bytes32 node
+    ) public {
         simpleMerkle.insert(node);
         fullMerkle.insert(node);
 
@@ -100,10 +102,10 @@ contract MerkleTest is Test {
         fullMerkle.insert(oldNode);
 
         bytes32[16] memory proof = fullMerkle.getProof(0);
-        
+
         simpleMerkle.update(newNode, oldNode, proof, 0);
         fullMerkle.update(newNode, 0);
-        
+
         // Verify new node
         proof = fullMerkle.getProof(0);
         assertTrue(simpleMerkle.verify(newNode, proof, 0));
